+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+- Added support to immediately execute a queue job from the admin control panel
 - Added a priority column to the queue jobs view in the admin control panel
 
 ### Changed